+++ KOP/KOP.WEB/Controllers/QualificationController.cs
@@ -11,19 +11,48 @@ namespace KOP.WEB.Controllers
     public class QualificationController : Controller
     {
         private readonly IGradeService _gradeService;
+        private readonly ILogger<QualificationController> _logger;
 
-        public QualificationController(IGradeService gradeService)
+        public QualificationController(IGradeService gradeService, ILogger<QualificationController> logger)
         {
             _gradeService = gradeService;
+            _logger = logger;
         }
 
         [HttpGet]
         [Authorize]
         public async Task<IActionResult> GetPopup(int gradeId)
         {
+            if (gradeId <= 0)
+            {
+                _logger.LogWarning("Invalid gradeId: {gradeId}", gradeId);
+                return BadRequest("Invalid grade ID.");
+            }
+
             try
             {
+                var selectedUserId = HttpContext.Session.GetInt32("SelectedUserId");
+
+                if (!selectedUserId.HasValue || selectedUserId <= 0)
+                {
+                    _logger.LogWarning("SelectedUserId is incorrect or not found in session.");
+                    return BadRequest("Selected user ID is not valid.");
+                }
+
+                var selectedUserFullName = HttpContext.Session.GetString("SelectedUserFullName");
+
+                if (selectedUserFullName == null)
+                {
+                    _logger.LogWarning("SelectedUserFullName is incorrect or not found in session.");
+                    return BadRequest("Selected user FullName is not valid.");
+                }
+
                 var gradeDto = await _gradeService.GetGradeDto(gradeId, new List<GradeEntities> { GradeEntities.Qualification });
+
+                if (gradeDto.QualificationDto == null)
+                {
+                    throw new Exception($"Qualification is null for Grade with ID {gradeId}.");
+                }
 
                 var conclusionEditAccess = User.IsInRole("Urp");
                 var editAccess = User.IsInRole("Urp");
@@ -32,9 +61,8 @@ namespace KOP.WEB.Controllers
                 var viewModel = new QualificationViewModel
                 {
                     GradeId = gradeId,
-                    SelectedUserFullName = HttpContext.Session.GetString("SelectedUserFullName") ?? "-",
-                    // CHTCK THIS !!!
-                    SelectedUserId = HttpContext.Session.GetInt32("SelectedUserId") ?? 0,
+                    SelectedUserFullName = selectedUserFullName,
+                    SelectedUserId = selectedUserId.Value,
                     Conclusion = gradeDto.QualificationConclusion,
                     Qualification = gradeDto.QualificationDto,
                     EditAccess = editAccess,
@@ -44,8 +72,9 @@ namespace KOP.WEB.Controllers
 
                 return View("_QualificationPartial", viewModel);
             }
-            catch
+            catch(Exception ex)
             {
+                _logger.LogError(ex, "[QualificationController.GetPopup] : ");
                 return View("Error", new ErrorViewModel
                 {
                     StatusCode = StatusCodes.InternalServerError,
@@ -58,19 +87,14 @@ namespace KOP.WEB.Controllers
         [Authorize]
         public async Task<IActionResult> Edit(QualificationViewModel viewModel)
         {
+            if (viewModel.GradeId <= 0)
+            {
+                _logger.LogWarning("Invalid gradeId: {gradeId}", viewModel.GradeId);
+                return BadRequest("Invalid grade ID.");
+            }
+
             try
             {
-                var editAccess = User.IsInRole("Urp");
-
-                if (!editAccess)
-                {
-                    return BadRequest(new
-                    {
-                        error = "Произошла ошибка при сохранении.",
-                        details = "Доступ запрещен",
-                    });
-                }
-
                 var gradeDto = await _gradeService.GetGradeDto(viewModel.GradeId, new List<GradeEntities> { GradeEntities.Qualification });
 
                 gradeDto.QualificationDto = viewModel.Qualification;
@@ -83,6 +107,7 @@ namespace KOP.WEB.Controllers
             }
             catch (Exception ex)
             {
+                _logger.LogError(ex, "[QualificationController.Edit] : ");
                 return BadRequest(new
                 {
                     error = "Произошла ошибка при сохранении.",
@@ -95,6 +120,12 @@ namespace KOP.WEB.Controllers
         [Authorize]
         public async Task<IActionResult> DeletePreviousJob(int id)
         {
+            if (id <= 0)
+            {
+                _logger.LogWarning("Invalid previousJobId: {id}", id);
+                return BadRequest("Invalid previousJob ID.");
+            }
+
             try
             {
                 await _gradeService.DeletePreviousJob(id);
@@ -103,6 +134,7 @@ namespace KOP.WEB.Controllers
             }
             catch (Exception ex)
             {
+                _logger.LogError(ex, "[QualificationController.DeletePreviousJob] : ");
                 return BadRequest(new
                 {
                     error = "Произошла ошибка при удалении.",
@@ -115,6 +147,12 @@ namespace KOP.WEB.Controllers
         [Authorize]
         public async Task<IActionResult> DeleteHigherEducation(int id)
         {
+            if (id <= 0)
+            {
+                _logger.LogWarning("Invalid higherEducationId: {id}", id);
+                return BadRequest("Invalid higherEducation ID.");
+            }
+
             try
             {
                 await _gradeService.DeleteHigherEducation(id);
@@ -123,6 +161,7 @@ namespace KOP.WEB.Controllers
             }
             catch (Exception ex)
             {
+                _logger.LogError(ex, "[QualificationController.DeleteHigherEducation] : ");
                 return BadRequest(new
                 {
                     error = "Произошла ошибка при удалении.",
